requirements.txt
accelerate==1.0.1
astunparse==1.6.3
brotlipy==0.7.0
conda-build==3.27.0
conda-content-trust==0.2.0
conda-libmamba-solver==23.7.0
datasets==2.15.0
expecttest==0.1.6
faiss-gpu==1.7.2
gmpy2==2.1.2
hypothesis==6.87.1
ipykernel==6.29.5
mkl-fft==1.3.8
mkl-random==1.2.4
networkx==3.1
optuna==4.0.0
pip==23.2.1
PySocks==1.7.1
rank-bm25==0.2.2
sentence-transformers==3.2.1
sympy==1.11.1
toolz==0.12.0
torchaudio==2.1.0
torchelastic==0.2.2
torchvision==0.16.0
triton==2.1.0
types-dataclasses==0.6.6
wandb==0.18.2
